RPM and DEB package releases for users
This change provides RPM and DEB package releases of ESLauncher2.  When helping users in eslauncher2 discord I found some to not be technically capable of launching the terminal and running eslauncher2.  For this reason, I am contributing standard installable software packages for ESLauncher2 to improve the user experience of non-technical users who want to play with development versions of Endless Sky.

[Learn more about FPM][fpm] (f'n package manager)

[fpm]: https://fpm.readthedocs.io/en/latest/index.html